+++ test/libsolidity/smtCheckerTests/operators/division_6.sol
@@ -1,13 +1,14 @@
 pragma experimental SMTChecker;
 contract C {
-	function mul(uint256 a, uint256 b) public pure returns (uint256) {
+	function mul(uint8 a, uint8 b) public pure returns (uint8) {
 		if (a == 0) {
 			return 0;
 		}
-		uint256 c = a * b;
+		uint8 c = a * b;
 		require(c / a == b);
 		return c;
 	}
 }
 // ----
-// Warning 4984: (160-165): CHC: Overflow (resulting value larger than 2**256 - 1) happens here.
+// Warning 4984: (152-157): CHC: Overflow (resulting value larger than 255) happens here.\nCounterexample:\n\na = 128\nb = 2\n = 0\n\nTransaction trace:\nconstructor()\nmul(128, 2)
+// Warning 6838: (169-179): BMC: Condition is always true.
